implement add qso. still need to implement configuration, ensure that db is open and created before add qso. display log in separate window

+++ database.h
@@ -1,10 +1,10 @@
 #ifndef DATABASE_H
 #define DATABASE_H
 
-
 #include <sqlite3.h>
 #include <QString>
 #include <QFile>
+#include <QDebug>
 
 
 class database
@@ -16,7 +16,7 @@ public:
     sqlite3 *db_open_connection(QString& dbname);
     void close_db(sqlite3* db);
     int execute_query(QString& sql, sqlite3 *db);
-    int insert_data(QString& dbname, QString& callsign, float freq,
+    int insert_data(QString& dbname, QString& callsign, double freq,
                     QString& mode, QString& date, QString& time,
                     int rst_s, int rst_r, QString& exchange);
